LayoutTests/imported/w3c/web-platform-tests/webxr/exclusive_requestFrame_nolayer.https-expected.txt


FAIL XRSession requestAnimationFrame must fail if the session has no baseLayer for immersive - webgl assert_equals: expected object "[object XRWebGLLayer]" but got object "[object XRWebGLLayer]"
FAIL XRSession requestAnimationFrame must fail if the session has no baseLayer for immersive - webgl2 assert_equals: expected object "[object XRWebGLLayer]" but got object "[object XRWebGLLayer]"
FAIL XRSession requestAnimationFrame must fail if the session has no baseLayer for non immersive - webgl assert_equals: expected object "[object XRWebGLLayer]" but got object "[object XRWebGLLayer]"
FAIL XRSession requestAnimationFrame must fail if the session has no baseLayer for non immersive - webgl2 assert_equals: expected object "[object XRWebGLLayer]" but got object "[object XRWebGLLayer]"
